backend -create user

+++ java-app/src/main/java/com/teamkeys/java_app/users/entity/UserEntity.java
@@ -29,6 +29,6 @@ public class UserEntity {
 	private String password;
 	
 	@Column(name = "phone")
-	private int phoneNumber;
+	private long phoneNumber;
 	
 }
